+++ Craftsman/Services/FileNames.cs
@@ -56,7 +56,7 @@ public static class FileNames
     public static string GetFunctionalFixtureName() => "FunctionalTestFixture";    
     public static string GetControllerName(string entityName) => $"{entityName}Controller";    
     public static string GetDatabaseEntityConfigName(string entityName) => $"{entityName}Configuration";    
-    public static string GetDatabaseHelperFileName() => $"DatabaseHelper";    
+    public static string GetMigrationHostedServiceFileName() => $"MigrationHostedService";    
     public static string GetEnvironmentServiceFileName() => $"EnvironmentService";    
     public static string GetEnvironmentServiceInterfaceFileName() => $"I{GetEnvironmentServiceFileName()}";
     public static string GetSeederName(Entity entity) => $"{entity.Name}Seeder";