+++ CHANGELOG.md
@@ -11,7 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
-* `DatabaseHelper` to automatically apply migrations when starting your project and apply any data seeding you may want
+* New `MigrationHostedService` to automatically apply migrations when starting your project
 * New `EnvironmentService` for easy access to environment variables
 
 ### Updates